% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_counter_data.R
\name{bind_counter_data}
\alias{bind_counter_data}
\title{A function that combines individual Logie counter data files.}
\usage{
bind_counter_data(path_to_folder, no_channels, site, year, max_pss,
  rows_rm = FALSE)
}
\arguments{
\item{path_to_folder}{The file path for the folder that contains all data files for processing.}

\item{no_channels}{The number of counter channels that were operated.}

\item{site}{Name of the study river.}

\item{year}{Year of counter operation.}

\item{max_pss}{The maximum pss size.}

\item{rows_rm}{Determines whether or not the rows removed from the data are returned Defaults to FALSE.}
}
\value{
A master data file of processed Logie counter data. The data are written to file as well as stored in the environment.
}
\description{
A function to combine and process Logie counter data.
}
