% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_signal_data.R
\name{bind_signal_data}
\alias{bind_signal_data}
\title{A function that combines individual Logie graphics data files.}
\usage{
bind_signal_data(path_to_folder, site, year, max_pss, rows_rm = FALSE)
}
\arguments{
\item{path_to_folder}{The file path for the folder that contains all data files for processing.}

\item{site}{Name of the study river.}

\item{year}{Year of counter operation.}

\item{max_pss}{The maximum pss size.}

\item{rows_rm}{Determines whether the rows removed from the data are returned as a .csv file in the working directory. Defaults to FALSE.}
}
\value{
A master data file of processed Logie signal data. The data are written to file as well as stored in the environment. All individual signal data files are combined into one master graphics file and written to the folder path as "all_signal_data.csv".
}
\description{
A function to combine and process Logie graphics data.
}
