% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_ghost_records.R
\name{ghosts}
\alias{ghosts}
\title{This function detects possible ghost traces in graphics validated excel sheets
it identifies traces with the same direction within 2 seconds of the previous trace if it has an equal or smaller pss on a different channel
A date.time column must be present in the excel file
The graphics validation column must be named graphics_validation
The excel file must have a "channel" and "pss" column}
\usage{
ghosts(data)
}
\description{
This function detects possible ghost traces in graphics validated excel sheets
it identifies traces with the same direction within 2 seconds of the previous trace if it has an equal or smaller pss on a different channel
A date.time column must be present in the excel file
The graphics validation column must be named graphics_validation
The excel file must have a "channel" and "pss" column
}
