% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_abundance.R
\name{plot_abundance}
\alias{plot_abundance}
\title{A function that plots daily up and down counts as well as cumulative counts of Logie counter data}
\usage{
plot_abundance(dataset, first_day, print_to_file = FALSE)
}
\arguments{
\item{dataset}{The dataset used to create the plots.}

\item{first_day}{The first day of the dataset you want to use. This parameter needs to be specified in year day format. Defaults to the first day in the dataset}

\item{print_to_file}{If TRUE, plot is saved to the working directory (defaults to FALSE).}
}
\value{
Two plots are created, one with daily up, down, and cumulative total counts, as well as one with ups and cumulative up counts only.
}
\description{
Plots daily up and down counts.
}
