% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_events.R
\name{plot_events}
\alias{plot_events}
\title{A function that plots events per hour of Logie counter data}
\usage{
plot_events(dataset, first_day = NULL, last_day = NULL,
  min_pss = NULL, max_pss = NULL, print_to_file = FALSE)
}
\arguments{
\item{dataset}{The dataset used to create the plots.}

\item{first_day}{The first day of the dataset you want to use. This parameter needs to be specified in year day format. Defaults to the first day in the dataset}

\item{last_day}{The last day of the dataset you want to use. This parameter needs to be specified in year day format. Defaults to the last day in the dataset.}

\item{min_pss}{The lower threshold PSS value to be plotted. Defaults to 0.}

\item{max_pss}{The upper threshold PSS value to be plotted. Defaults to 130.}

\item{print_to_file}{If TRUE, plot is saved to the working directory (defaults to FALSE).}
}
\value{
Generates three histograms of peak signal size for up counts, down counts and events for each counter channel.
}
\description{
This function plots the number of events per hour, channel, and day.
}
