% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pss_hour.R
\name{plot_pss_hour}
\alias{plot_pss_hour}
\title{A function that plots average hourly peak signal size (pss) for Logie counter data}
\usage{
plot_pss_hour(dataset, description, first_day = NULL, last_day = NULL,
  min_pss = NULL, max_pss = NULL, ch = NULL, print_to_file = FALSE)
}
\arguments{
\item{dataset}{The cleaned counter dataset used to populate histograms (i.e., counter_data as created by bind_counter_data()).}

\item{description}{The type of counter data to be plotted. Must be "U" (ups), "D" (downs), or "E" (events).}

\item{first_day}{The first day of the dataset to be plotted, which must be specified in year-day format. Defaults to the first day in the dataset.}

\item{last_day}{The last day of the dataset to be plotted, which must be specified in year-day format. Defaults to the last day in the dataset.}

\item{min_pss}{The lower threshold peak signal size (pss) value to be plotted. Defaults to 0.}

\item{max_pss}{The upper threshold peak signal size (pss) value to be plotted. Defaults to 130.}

\item{ch}{The channel to be plotted. Defaults to all channels. Needs to be inputed as an object or vector (e.g. 1, or c(1, 2)).}

\item{print_to_file}{If TRUE, plot is saved to the working directory. Defaults to FALSE.}
}
\value{
Generates two plots of average hourly pss data. One plot displays a time series of average hourly pss for all channels combined (as specified by ch). 
A second plot displayes a time series of average hourly pss separated into each channel specified by ch.
}
\description{
This function plots the average hourly peak signal size (pss) for Logie counter data.
}
