% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_records.R
\name{hist_records}
\alias{hist_records}
\title{A function that plots histograms of Logie counter data for the specified description}
\usage{
hist_records(dataset, description, first_day = NULL, last_day = NULL,
  min_pss = NULL, max_pss = NULL, print_to_file = FALSE)
}
\arguments{
\item{dataset}{The dataset used to create the histograms.}

\item{description}{The type of counter event to be plotted. Must be "U", "D", or "E".}

\item{first_day}{The first day of the dataset you want to use. This parameter needs to be specified in year day format. Defaults to the first day in the dataset}

\item{last_day}{The last day of the dataset you want to use. This parameter needs to be specified in year day format. Defaults to the last day in the dataset.}

\item{min_pss}{The lower threshold PSS value to be plotted. Defaults to 0.}

\item{max_pss}{The upper threshold PSS value to be plotted. Defaults to 130.}

\item{print_to_file}{If TRUE, plot is saved to the working directory (defaults to FALSE).}
}
\value{
Generates a histogram of peak signal size for either up counts, down counts, or events for each counter channel. Also prints a summary of the total number of events per channel for the specified description.
}
\description{
This function plots historgrams of up, down, and event counts for Logie counter data by channel as specified by the user.
}
