% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pss_hour.R
\name{plot_pss_hour}
\alias{plot_pss_hour}
\title{A function that plots the peak signal size (PSS) by hour of Logie counter data}
\usage{
plot_pss_hour(dataset, description, first_day = NULL, last_day = NULL,
  min_pss = NULL, max_pss = NULL, ch = NULL, print_to_file = FALSE)
}
\arguments{
\item{dataset}{The dataset used to create the plots.}

\item{description}{The type of counter event to be plotted. Must be "U", "D", or "E".}

\item{first_day}{The first day of the dataset you want to use. This parameter needs to be specified in year day format. Defaults to the first day in the dataset}

\item{last_day}{The last day of the dataset you want to use. This parameter needs to be specified in year day format. Defaults to the last day in the dataset.}

\item{min_pss}{The lower threshold PSS value to be plotted. Defaults to 0.}

\item{max_pss}{The upper threshold PSS value to be plotted. Defaults to 130.}

\item{ch}{The channel to be plotted. Defaults to all channels. Needs to be inputed as an object or vector (e.g. 1, or c(1, 2))}

\item{print_to_file}{If TRUE, plot is saved to the working directory (defaults to FALSE).}
}
\value{
Generates a plot of total up counts per hour and peak pss size by hour.
}
\description{
This function plots the PSS by hour for Logie counter data
}
